requirements.txt
numpy>=1.24.0
scipy>=1.10.0
moviepy>=1.0.3
pydub>=0.25.1
matplotlib>=3.7.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
